requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.104.1
filelock==3.9.0
fsspec==2023.4.0
h11==0.14.0
httptools==0.6.1
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
Pillow==9.3.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.28.1
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
typing_extensions==4.8.0
urllib3==1.26.13
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0